requirements.txt
fastapi==0.115.6
numpy==2.2.1
openai==1.58.1
pymongo==4.6.1
PyPDF2==3.0.1
python-dotenv==1.0.1
Requests==2.32.3
uvicorn==0.34.0
